pyproject.toml
[project]
name = "painel-secom-vivi"
version = "0.1.0"
description = ""
authors = [
    {name = "Viviane Nascimento",email = "vivigoncalvesporto@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "frictionless (>=5.18.1,<6.0.0)",
    "datapackage (>=1.15.4,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
